fix(collab): address CodeRabbit review on chat + dialog mounting (#754)
- Track unread/auto-scroll by the latest message id instead of chat.length:
  the in-store chat is a bounded tail, so once full new messages arrive
  without changing the length and would stop updating the count/scroll. Seed
  the baseline on first open so welcome history isn't marked unread.
- Don't clear the chat composer (or allow send) while reconnecting: sendChat
  silently no-ops when the socket isn't open, so clearing would lose the text.
- Move the CollaborateDialog render and the ?collab= auto-open effect from
  TopToolbar to DesktopShell, so the dialog the on-canvas badge reopens stays
  mounted in toolbar-hidden layouts. TopToolbar keeps only the trigger.
